requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
idna==3.4
msal==1.24.1
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
urllib3==2.0.7